requirements.txt
fastai
opencv-python
numpy==1.21.5
Pillow==9.0.0
telebot
python-telegram-bot==13.8